@@ -1,9 +1,14 @@
+from __future__ import annotations
+
 import logging
 from hashlib import md5
 from pathlib import Path
-from typing import Literal, Sequence, Union
+from typing import TYPE_CHECKING, Literal
 
 import pytest
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
 
 logger = logging.getLogger("atomate2")
 
@@ -126,7 +131,7 @@ def mock_cp2k(monkeypatch, cp2k_test_dir):
 
 
 def fake_run_cp2k(
-    ref_path: Union[str, Path],
+    ref_path: str | Path,
     input_settings: Sequence[str] = (),
     check_inputs: Sequence[Literal["cp2k.inp"]] = _VFILES,
     clear_inputs: bool = True,
@@ -170,8 +175,8 @@ def check_input():
         from pymatgen.io.cp2k.inputs import Cp2kInput
 
         ref = Cp2kInput.from_file(ref_path / "inputs" / "cp2k.inp")
-        user_input.verbosity(False)
-        ref.verbosity(False)
+        user_input.verbosity(verbosity=False)
+        ref.verbosity(verbosity=False)
         user_string = " ".join(user_input.get_string().lower().split())
         user_hash = md5(user_string.encode("utf-8")).hexdigest()
 
@@ -194,7 +199,7 @@ def clear_cp2k_inputs():
     logger.info("Cleared cp2k inputs")
 
 
-def copy_cp2k_outputs(ref_path: Union[str, Path]):
+def copy_cp2k_outputs(ref_path: str | Path):
     import shutil
 
     output_path = ref_path / "outputs"
