@@ -6,7 +6,7 @@ from atomate2.cli.dev import dev
 
 
 @click.group(context_settings={"help_option_names": ["-h", "--help"]})
-def cli():
+def cli() -> None:
     """Command-line interface for atomate2."""
 
 
