@@ -4,13 +4,13 @@ import click
 
 
 @click.group(context_settings={"help_option_names": ["-h", "--help"]})
-def dev():
+def dev() -> None:
     """Tools for atomate2 developers."""
 
 
 @dev.command(context_settings={"help_option_names": ["-h", "--help"]})
 @click.argument("test_dir")
-def vasp_test_data(test_dir):
+def vasp_test_data(test_dir) -> None:
     """Generate test data for VASP unit tests.
 
     This script expects there is an outputs.json file and job folders in the current
@@ -158,7 +158,7 @@ def test_my_flow(mock_vasp, clean_dir, si_structure):
     print(test_function_str)
 
 
-def _potcar_to_potcar_spec(potcar_filename, output_filename):
+def _potcar_to_potcar_spec(potcar_filename, output_filename) -> None:
     """Convert a POTCAR file to a POTCAR.spec file."""
     from pymatgen.io.vasp import Potcar
 
