@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, Callable, Iterable
+from typing import TYPE_CHECKING, Callable
 
 import numpy as np
 from jobflow import Flow, Response, job
@@ -19,6 +19,7 @@ from pymatgen.entries.computed_entries import ComputedStructureEntry
 from atomate2.common.schemas.defects import CCDDocument
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
     from pathlib import Path
 
     from emmet.core.tasks import TaskDoc
@@ -40,7 +41,7 @@ class CCDInput(BaseModel):
 
 
 @job
-def get_charged_structures(structure: Structure, charges: Iterable):
+def get_charged_structures(structure: Structure, charges: Iterable) -> list[Structure]:
     """Add charges to a structure.
 
     This needs to be a job so the results of other jobs can be passed in.
@@ -72,7 +73,7 @@ def spawn_energy_curve_calcs(
     prev_vasp_dir: str | Path | None = None,
     add_name: str = "",
     add_info: dict | None = None,
-):
+) -> Response:
     """Compute the total energy curve from a reference to distorted structure.
 
     Parameters
@@ -144,7 +145,7 @@ def get_ccd_documents(
     inputs1: Iterable[CCDInput],
     inputs2: Iterable[CCDInput],
     undistorted_index: int,
-):
+) -> Response:
     """
     Get the configuration coordinate diagram from the task documents.
 
@@ -406,7 +407,7 @@ def check_charge_state(charge_state: int, task_structure: Structure) -> Response
 
 
 @job
-def get_defect_entry(charge_state_summary: dict, bulk_summary: dict):
+def get_defect_entry(charge_state_summary: dict, bulk_summary: dict) -> list[dict]:
     """Get a defect entry from a defect calculation and a bulk calculation."""
     bulk_sc_entry = bulk_summary["sc_entry"]
     bulk_struct_entry = ComputedStructureEntry(
