@@ -36,7 +36,7 @@ logger = logging.getLogger(__name__)
 @job
 def get_total_energy_per_cell(
     total_dft_energy_per_formula_unit: float, structure: Structure
-):
+) -> float:
     """
     Job that computes total dft energy of the cell.
 
@@ -58,7 +58,7 @@ def get_total_energy_per_cell(
 @job
 def get_supercell_size(
     structure: Structure, min_length: float, prefer_90_degrees: bool, **kwargs
-):
+) -> list[list[float]]:
     """
     Determine supercell size with given min_length.
 
@@ -125,7 +125,7 @@ def generate_phonon_displacements(
     use_symmetrized_structure: str | None,
     kpath_scheme: str,
     code: str,
-):
+) -> list[Structure]:
     """
     Generate displaced structures with phonopy.
 
@@ -195,7 +195,7 @@ def generate_frequencies_eigenvectors(
     epsilon_static: Matrix3D = None,
     born: Matrix3D = None,
     **kwargs,
-):
+) -> PhononBSDOSDoc:
     """
     Analyze the phonon runs and summarize the results.
 
@@ -254,7 +254,7 @@ def run_phonon_displacements(
     prev_dir: str | Path = None,
     prev_dir_argname: str = None,
     socket: bool = False,
-):
+) -> Flow:
     """
     Run phonon displacements.
 
