@@ -23,7 +23,7 @@ class Cp2kDrone(AbstractDrone):
         Additional keyword args passed to :obj:`.TaskDocument.from_directory`.
     """
 
-    def __init__(self, **task_document_kwargs):
+    def __init__(self, **task_document_kwargs) -> None:
         self.task_document_kwargs = task_document_kwargs
 
     def assimilate(self, path: str | Path | None = None) -> TaskDocument:
