@@ -46,7 +46,9 @@ class DoubleRelaxMaker(Maker):
     relax_maker1: Maker = field(default_factory=RelaxMaker)
     relax_maker2: Maker = field(default_factory=RelaxMaker)
 
-    def make(self, structure: Structure, prev_cp2k_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_cp2k_dir: str | Path | None = None
+    ) -> Flow:
         """
         Create a flow with two chained relaxations.
 
@@ -73,7 +75,7 @@ class DoubleRelaxMaker(Maker):
         return Flow([relax1, relax2], relax2.output, name=self.name)
 
     @classmethod
-    def from_relax_maker(cls, relax_maker: BaseCp2kMaker):
+    def from_relax_maker(cls, relax_maker: BaseCp2kMaker) -> DoubleRelaxMaker:
         """
         Instantiate the DoubleRelaxMaker with two relax makers of the same type.
 
@@ -112,7 +114,9 @@ class BandStructureMaker(Maker):
     static_maker: Maker = field(default_factory=StaticMaker)
     bs_maker: Maker = field(default_factory=NonSCFMaker)
 
-    def make(self, structure: Structure, prev_cp2k_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_cp2k_dir: str | Path | None = None
+    ) -> Flow:
         """
         Create a band structure flow.
 
@@ -189,7 +193,9 @@ class RelaxBandStructureMaker(Maker):
     relax_maker: Maker = field(default_factory=DoubleRelaxMaker)
     band_structure_maker: Maker = field(default_factory=BandStructureMaker)
 
-    def make(self, structure: Structure, prev_cp2k_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_cp2k_dir: str | Path | None = None
+    ) -> Flow:
         """
         Run a relaxation and then calculate the uniform and line mode band structures.
 
@@ -238,7 +244,7 @@ class HybridFlowMaker(Maker):
     pbe_maker: Maker = field(default=lambda: StaticMaker(store_output_data=False))
     hybrid_maker: Maker = field(default_factory=HybridStaticMaker)
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         """
         Post init updates.
 
@@ -249,7 +255,9 @@ class HybridFlowMaker(Maker):
         the PBE density matrix, which creates huge speed-ups. Rarely causes
         problems so it is done as a default here.
         """
-        updates = {"activate_hybrid": {"hybrid_functional": self.hybrid_functional}}
+        updates: dict[str, dict[str, str | bool]] = {
+            "activate_hybrid": {"hybrid_functional": self.hybrid_functional}
+        }
         if self.initialize_with_pbe:
             updates["activate_hybrid"].update(
                 {"screen_on_initial_p": True, "screen_p_forces": True}
