@@ -302,7 +302,7 @@ class NonSCFMaker(BaseCp2kMaker):
         structure: Structure,
         prev_cp2k_dir: str | Path | None,
         mode: str = "uniform",
-    ):
+    ) -> None:
         """
         Run a non-scf CP2K job.
 
@@ -377,7 +377,7 @@ class TransmuterMaker(BaseCp2kMaker):
         self,
         structure: Structure,
         prev_cp2k_dir: str | Path | None = None,
-    ):
+    ) -> None:
         """
         Run a transmuter Cp2k job.
 
