@@ -6,7 +6,7 @@ import logging
 import shlex
 import subprocess
 from os.path import expandvars
-from typing import TYPE_CHECKING, Any, Sequence
+from typing import TYPE_CHECKING, Any
 
 from custodian import Custodian
 from custodian.cp2k.handlers import (
@@ -26,6 +26,8 @@ from jobflow.utils import ValueEnum
 from atomate2 import SETTINGS
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     from custodian.custodian import ErrorHandler, Validator
 
     from atomate2.cp2k.schemas.task import TaskDocument
@@ -66,7 +68,7 @@ def run_cp2k(
     validators: Sequence[Validator] = _DEFAULT_VALIDATORS,
     cp2k_job_kwargs: dict[str, Any] = None,
     custodian_kwargs: dict[str, Any] = None,
-):
+) -> None:
     """
     Run CP2K.
 
