@@ -40,7 +40,7 @@ class Cp2kInputSet(InputSet):
         self,
         cp2k_input: Cp2kInput,
         optional_files: dict | None = None,
-    ):
+    ) -> None:
         """
         Initialize the set.
 
@@ -84,7 +84,7 @@ class Cp2kInputSet(InputSet):
         directory: str | Path,
         make_dir: bool = True,
         overwrite: bool = True,
-    ):
+    ) -> None:
         """
         Write Cp2k input file to a directory.
 
@@ -116,7 +116,9 @@ class Cp2kInputSet(InputSet):
                 raise FileExistsError(f"{directory / fn} already exists.")
 
     @staticmethod
-    def from_directory(directory: str | Path, optional_files: dict = None):
+    def from_directory(
+        directory: str | Path, optional_files: dict = None
+    ) -> Cp2kInputSet:
         """
         Load a set of CP2K inputs from a directory.
 
@@ -188,7 +190,7 @@ class Cp2kInputGenerator(InputGenerator):
     force_gamma: bool = False
     config_dict: dict = field(default_factory=lambda: _BASE_CP2K_SET)
 
-    def get_input_set(  # type: ignore
+    def get_input_set(
         self,
         structure: Structure | Molecule = None,
         prev_dir: str | Path = None,
@@ -247,7 +249,7 @@ class Cp2kInputGenerator(InputGenerator):
 
         return Cp2kInputSet(cp2k_input=cp2k_input, optional_files=optional_files)
 
-    def get_input_updates(self, structure, prev_input) -> dict:
+    def get_input_updates(self, structure: Structure, prev_input: Cp2kInput) -> dict:
         """
         Get updates to the cp2k input for this calculation type.
 
@@ -295,7 +297,9 @@ class Cp2kInputGenerator(InputGenerator):
         """
         return {}
 
-    def _get_previous(self, structure: Structure = None, prev_dir: str | Path = None):
+    def _get_previous(
+        self, structure: Structure = None, prev_dir: str | Path = None
+    ) -> tuple[Structure, Cp2kInput, Cp2kOutput]:
         """Load previous calculation outputs and decide which structure to use."""
         if structure is None and prev_dir is None:
             raise ValueError("Either structure or prev_dir must be set.")
@@ -313,7 +317,7 @@ class Cp2kInputGenerator(InputGenerator):
 
         return structure, prev_input, cp2k_output
 
-    def _get_structure(self, structure):
+    def _get_structure(self, structure) -> Structure:
         """Get the standardized structure."""
         if self.sort_structure and hasattr(structure, "get_sorted_structure"):
             structure = structure.get_sorted_structure()
@@ -325,7 +329,7 @@ class Cp2kInputGenerator(InputGenerator):
         kpoints: Kpoints | None = None,
         previous_input: Cp2kInput = None,
         input_updates: dict = None,
-    ):
+    ) -> Cp2kInput:
         """Get the input."""
         previous_input = previous_input or {}
         input_updates = input_updates or {}
@@ -355,7 +359,7 @@ class Cp2kInputGenerator(InputGenerator):
         cp2k_input.update(overrides)
         return cp2k_input
 
-    def _get_basis_file(self, cp2k_input: Cp2kInput):
+    def _get_basis_file(self, cp2k_input: Cp2kInput) -> BasisFile:
         """
         Get the basis sets for the input object and convert them to a basis file object.
 
@@ -372,7 +376,7 @@ class Cp2kInputGenerator(InputGenerator):
         cp2k_input.safeset({"force_eval": {"dft": {"BASIS_SET_FILE_NAME": "BASIS"}}})
         return BasisFile(objects=basis_sets)
 
-    def _get_potential_file(self, cp2k_input: Cp2kInput):
+    def _get_potential_file(self, cp2k_input: Cp2kInput) -> PotentialFile:
         """
         Get the potentials and convert them to a potential file object.
 
@@ -531,7 +535,7 @@ class Cp2kInputGenerator(InputGenerator):
 # TODO From `atomate2.vasp.sets.base`. Should possibly go in common.
 # only reservation is if, eventually, CP2K gets it own kpoint object version
 # instead of using the vasp kpoint objects.
-def _combine_kpoints(*kpoints_objects: Kpoints):
+def _combine_kpoints(*kpoints_objects: Kpoints) -> Kpoints:
     """Combine k-points files together."""
     labels = []
     kpoints = []
@@ -595,7 +599,7 @@ class Cp2kAllElectronInputGenerator(Cp2kInputGenerator):
         return None
 
 
-def recursive_update(d: dict, u: dict):
+def recursive_update(d: dict, u: dict) -> dict:
     """
     Update a dictionary recursively and return it.
 
