@@ -35,7 +35,7 @@ class CHGNetVaspRelaxMaker(Maker):
     chgnet_maker: CHGNetRelaxMaker = field(default_factory=CHGNetRelaxMaker)
     vasp_maker: BaseVaspMaker = field(default_factory=RelaxMaker)
 
-    def make(self, structure: Structure):
+    def make(self, structure: Structure) -> Flow:
         """
         Create a flow with a CHGNet (pre)relaxation followed by a VASP relaxation.
 
@@ -76,7 +76,7 @@ class M3GNetVaspRelaxMaker(Maker):
     m3gnet_maker: M3GNetRelaxMaker = field(default_factory=M3GNetRelaxMaker)
     vasp_maker: BaseVaspMaker = field(default_factory=RelaxMaker)
 
-    def make(self, structure: Structure):
+    def make(self, structure: Structure) -> Flow:
         """
         Create a flow with a M3GNet (pre)relaxation followed by a VASP relaxation.
 
