@@ -51,7 +51,7 @@ class ForceFieldRelaxMaker(Maker):
     task_document_kwargs: dict = field(default_factory=dict)
 
     @job(output_schema=ForceFieldTaskDocument)
-    def make(self, structure: Structure):
+    def make(self, structure: Structure) -> ForceFieldTaskDocument:
         """
         Perform a relaxation of a structure using a force field.
 
@@ -78,7 +78,7 @@ class ForceFieldRelaxMaker(Maker):
             **self.task_document_kwargs,
         )
 
-    def _relax(self, structure):
+    def _relax(self, structure: Structure) -> dict:
         raise NotImplementedError
 
 
@@ -102,7 +102,7 @@ class ForceFieldStaticMaker(ForceFieldRelaxMaker):
     task_document_kwargs: dict = field(default_factory=dict)
 
     @job(output_schema=ForceFieldTaskDocument)
-    def make(self, structure: Structure):
+    def make(self, structure: Structure) -> ForceFieldTaskDocument:
         """
         Perform a static evaluation using a force field.
 
@@ -122,14 +122,14 @@ class ForceFieldStaticMaker(ForceFieldRelaxMaker):
         return ForceFieldTaskDocument.from_ase_compatible_result(
             self.force_field_name,
             result,
-            False,
-            1,
-            None,
-            None,
+            relax_cell=False,
+            steps=1,
+            relax_kwargs=None,
+            optimizer_kwargs=None,
             **self.task_document_kwargs,
         )
 
-    def _evaluate_static(self, structure):
+    def _evaluate_static(self, structure: Structure) -> dict:
         raise NotImplementedError
 
 
@@ -162,7 +162,7 @@ class CHGNetRelaxMaker(ForceFieldRelaxMaker):
     optimizer_kwargs: dict = field(default_factory=dict)
     task_document_kwargs: dict = field(default_factory=dict)
 
-    def _relax(self, structure):
+    def _relax(self, structure: Structure) -> dict:
         from chgnet.model import StructOptimizer
 
         relaxer = StructOptimizer(**self.optimizer_kwargs)
@@ -188,7 +188,7 @@ class CHGNetStaticMaker(ForceFieldStaticMaker):
     force_field_name = "CHGNet"
     task_document_kwargs: dict = field(default_factory=dict)
 
-    def _evaluate_static(self, structure):
+    def _evaluate_static(self, structure: Structure) -> dict:
         from chgnet.model import StructOptimizer
 
         relaxer = StructOptimizer()
@@ -226,7 +226,7 @@ class M3GNetRelaxMaker(ForceFieldRelaxMaker):
     optimizer_kwargs: dict = field(default_factory=dict)
     task_document_kwargs: dict = field(default_factory=dict)
 
-    def _relax(self, structure):
+    def _relax(self, structure: Structure) -> dict:
         import matgl
         from matgl.ext.ase import Relaxer
 
@@ -266,7 +266,7 @@ class M3GNetStaticMaker(ForceFieldStaticMaker):
     force_field_name: str = "M3GNet"
     task_document_kwargs: dict = field(default_factory=dict)
 
-    def _evaluate_static(self, structure):
+    def _evaluate_static(self, structure: Structure) -> dict:
         import matgl
         from matgl.ext.ase import Relaxer
 
@@ -325,7 +325,7 @@ class GAPRelaxMaker(ForceFieldRelaxMaker):
     potential_param_file_name: str = "gap.xml"
     potential_kwargs: dict = field(default_factory=dict)
 
-    def _relax(self, structure):
+    def _relax(self, structure: Structure) -> dict:
         from quippy.potential import Potential
 
         calculator = Potential(
@@ -364,7 +364,7 @@ class GAPStaticMaker(ForceFieldStaticMaker):
     potential_param_file_name: str | Path = "gap.xml"
     potential_kwargs: dict = field(default_factory=dict)
 
-    def _evaluate_static(self, structure):
+    def _evaluate_static(self, structure: Structure) -> dict:
         from quippy.potential import Potential
 
         calculator = Potential(
