@@ -35,6 +35,7 @@ from atomate2.vasp.sets.core import (
 if TYPE_CHECKING:
     from pathlib import Path
 
+    from jobflow import Response
     from pymatgen.core.structure import Structure
 
     from atomate2.vasp.sets.base import VaspInputGenerator
@@ -182,7 +183,7 @@ class NonSCFMaker(BaseVaspMaker):
         structure: Structure,
         prev_vasp_dir: str | Path | None,
         mode: str = "uniform",
-    ):
+    ) -> Response:
         """
         Run a non-scf VASP job.
 
@@ -356,7 +357,7 @@ class HSEBSMaker(BaseVaspMaker):
         structure: Structure,
         prev_vasp_dir: str | Path | None = None,
         mode="uniform",
-    ):
+    ) -> Response:
         """
         Run a HSE06 band structure VASP job.
 
@@ -487,7 +488,7 @@ class TransmuterMaker(BaseVaspMaker):
         self,
         structure: Structure,
         prev_vasp_dir: str | Path | None = None,
-    ):
+    ) -> Response:
         """
         Run a transmuter VASP job.
 
